requirements.txt
openai>=1.51.0
pyserini>=0.24.0
pydantic>=2.9.0
requests>=2.31.0
python-dotenv>=1.0.1

